Workaround to use correct WebInspector resource paths
The required WebInspector resources are shipped in `libWPEWebInspectorResources.so`,
but the WebInspectorServer incorrectly assumes that all the resource paths
are `/org/wpe/inspector/Userinterface`. In fact, the actual interface
is shipped with a different path.
```
...
InspectorGResourceBundle   2514 c /org/webkitgtk/inspector/UserInterface/Main.html
WebKit2InspectorGResourceBundle   1395 u /org/wpe/inspector/UserInterface/inspectorPageIndex.html
```

This workaround corrects the workaround for the case when WPE is
compiled with pretty basic settings in `meta-wpe`, by hardcoding
the relevant paths for that resource compilation.

Signed-off-by: Gergely Imreh <imrehg@gmail.com>